Ensure btree comparison doesn't fail for smallish N

(Which is to say, differences between values exceeding the max value
of int.)

+++ enterprise/b/btree.go
@@ -75,7 +75,7 @@ type (
 	//	  0 if a == b
 	//	> 0 if a >  b
 	//
-	Cmp func(a, b uint64) int
+	Cmp func(a, b uint64) int64
 
 	d struct { // data page
 		c int
